mtce

+++ applications/genHigherClass_jenkins/generate_jenkins.php
@@ -54,10 +54,18 @@ if(pathinfo($file, PATHINFO_EXTENSION) == "zip")
 else
 {
     /* this will zip the uploaded file then delete the uploaded file then unzip it. This is so that Jenkins will own the file to be processed */
-    shell_exec("zip $file.zip $file");
-    unlink($file);
-    shell_exec("unzip $file.zip");
-    unlink("$file.zip");
+    if(file_exists($file)) {
+        echo "\nOK: File exists: $file\n";
+        echo "\n[zip $file.zip $file]\n";
+        shell_exec("zip $file.zip $file");
+        echo "\n[$file]\n";
+        unlink($file);
+        shell_exec("unzip $file.zip");
+        unlink("$file.zip");
+    }
+    else echo "\nERROR: File does not exist: $file\n";
+    
+    
 }
 
 require_library('connectors/DwCA_Utility_cmd');
